machine-learning/model-building/requirements.txt
apyori==1.1.2
backcall==0.2.0
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
entrypoints==0.3
greenlet==1.1.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
jupyter-client==7.0.3
jupyter-core==4.8.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.2
pandas==1.3.3
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
ptyprocess==0.7.0
pycausalimpact==0.1.1
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.3.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.23
statsmodels==0.13.0
threadpoolctl==2.2.0
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5
